+++ doc/Changelog.md
@@ -4,6 +4,8 @@
 
 **In development**
 
+- [PR118](https://github.com/RoseauTechnologies/Roseau_Load_Flow/pull/118) [GH95](https://github.com/RoseauTechnologies/Roseau_Load_Flow/issues/95)
+  Add a catalogue of three-phase MV/LV distribution transformers
 - [PR117](https://github.com/RoseauTechnologies/Roseau_Load_Flow/pull/117) Add prettier to pre-commit
 - [PR116](https://github.com/RoseauTechnologies/Roseau_Load_Flow/pull/116)
   - Add a catalogue of networks